+++ collective/behavior/relatedmedia/interfaces.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from collective.behavior.relatedmedia import messageFactory as _
 from zope import schema
 from zope.interface import Interface
@@ -25,7 +24,7 @@ class IRelatedMediaSettings(Interface):
         title=_("Create Media Container in Assets Folder " "(language independent)?"),
         description=_(
             "If True, the Media Container path defined above is "
-            "generated in the language independend Assets folder. "
+            "generated in the language independent Assets folder. "
             "This requires plone.app.multilingual."
         ),
         default=False,
@@ -72,6 +71,12 @@ class IRelatedMediaSettings(Interface):
     show_titles_as_caption_default = schema.Bool(
         title=_("Show image titles as caption"),
         default=False,
+        required=False,
+    )
+
+    show_images_viewlet_default = schema.Bool(
+        title=_("Show images in viewlet"),
+        default=True,
         required=False,
     )