+++ collective/behavior/relatedmedia/widget.py
@@ -1,6 +1,4 @@
-# -*- coding: utf-8 -*-
 from plone.app.z3cform.widgets.relateditems import RelatedItemsWidget
-from plone.base import PloneMessageFactory as _
 from z3c.form.interfaces import IFieldWidget
 from z3c.form.interfaces import ITextWidget
 from z3c.form.widget import FieldWidget