+++ resources/bundle.js
@@ -1,0 +1,6 @@
+import registry from "@patternslib/patternslib/src/core/registry";
+
+
+import "./pat-related-images/related-images";
+
+registry.init();